Rename what broke.
